updated Gebindeproblem

+++ adapters/src/main/java/de/dhbw/foodcoop/warehouse/adapters/representations/BrotBestandRepresentation.java
@@ -2,18 +2,18 @@ package de.dhbw.foodcoop.warehouse.adapters.representations;
 
 public class BrotBestandRepresentation extends BestandRepresentation{
 
-    private long gewicht;
+    private double gewicht;
     
 	
 	
-	public BrotBestandRepresentation(String id, String name, boolean verfuegbarkeit, float preis, long gewicht) {
+	public BrotBestandRepresentation(String id, String name, boolean verfuegbarkeit, float preis, double gewicht) {
 		super(id, name, verfuegbarkeit, preis);
 		this.gewicht = gewicht;
 	}
 
 
 
-	public long getGewicht() {
+	public double getGewicht() {
 		return gewicht;
 	}
 }
